separate legacy tests for now
* separate legacy tests for now

we can find a better pattern for organizing the tests into different files and breaking the tests further down to cover more units
also some minor cleanup

* start grouping new tests into top level parser test files
