+++ src/__tests__/parser/legacy.test.ts
@@ -7,96 +7,7 @@ import { parse, withCustomConfig, withDefaultConfig } from "../../parser";
 import { getDefaultExportForFile } from "../../parser/utilities";
 import type { PropFilter } from "../../parser/types";
 
-describe.concurrent("parser", () => {
-  it("parses plain function", () => {
-    const result = parse(fixturePath("plain/function"));
-
-    expect(result).toEqual([
-      {
-        displayName: "add",
-        params: [
-          {
-            description: "the magical first number",
-            name: "a",
-            type: "number",
-          },
-          {
-            description: "boring second number",
-            name: "b",
-            type: "number",
-          },
-        ],
-        description: "get the sum of 2 numbers",
-        returns: "string",
-        tags: {
-          param: "a the magical first number\nb boring second number",
-        },
-        methods: [],
-        props: {},
-        filePath: fixturePath("plain/function"),
-      },
-    ]);
-  });
-
-  it("parses multiple plain functions but ignores those without descriptions (for backwards compatibility)", () => {
-    const result = parse(fixturePath("plain/functions"));
-
-    expect(result).toEqual([
-      {
-        displayName: "add",
-        params: [
-          {
-            description: "the magical first number",
-            name: "a",
-            type: "number",
-          },
-          {
-            description: "boring second number",
-            name: "b",
-            type: "number",
-          },
-        ],
-        description: "get the sum of 2 numbers",
-        returns: "string",
-        tags: {
-          param: "a the magical first number\nb boring second number",
-        },
-        methods: [],
-        props: {},
-        filePath: fixturePath("plain/functions"),
-      },
-    ]);
-  });
-
-  it("parses plain function with rest parameter", () => {
-    const filePath = fixturePath("plain/functionWithRestParam");
-
-    const result = parse(filePath);
-
-    expect(result).toEqual([
-      {
-        displayName: "sum",
-        params: [{ name: "numbers", type: "number[]", description: "" }],
-        filePath,
-        methods: [],
-        tags: {
-          param: "numbers",
-        },
-        props: {},
-        returns: "number",
-        description: "Gets the sum of all numbers passed in",
-      },
-    ]);
-  });
-
-  it("should parse functional component with spread props", () => {
-    check("FunctionalComponentWithSpreadProps", {
-      Jumbotron: {
-        prop1: { type: "string", required: true },
-      },
-    });
-  });
-
+describe.concurrent("parser legacy integration tests", () => {
   it("should parse simple react class component", () => {
     check("Column", {
       Column: {
